// src/utils/enums/test.case.data.ts
export enum TEST_CASE_DATA {
    NAME = 'Shehryar',
    EMAIL = 'shehryar_tanoli@outlook.com',
    PASSWORD = 'password@P123',
    NEW_PASSWORD = 'newpassword@P123',
    NEW_EMAIL = 'Asfandyar@outlook.com',
    ID_1 = '1',
    ID_2 = '2',
    UUID= '550e8400-e29b-41d4-a716-446655440000',
    ANOTHER_UUID = '550e8400-e29b-41d4-a716-446655440001',
    LOGGED_IN_USER_EMAIL = 'loggedInUser@outlook.com',
    EXISTING_EMAIL ='existing@outlook.com'
  }
  